+++ packages/modules/sentry/src/sentry.ts
@@ -5,25 +5,8 @@ import { Event } from '@sentry/types'
 
 import { serializeEnvelope } from './serialize-envelope'
 
-export const getCookies = (cookie: string): Record<string, string> =>
-  Object.fromEntries(
-    cookie
-      .split('; ')
-      .map((v) => {
-        try {
-          return v.split(/=(.*)/s).map(decodeURIComponent)
-        } catch {
-          return ['', '']
-        }
-      })
-      .filter(([v]) => !!v),
-  )
-
 export const sendEnvelope = async (sentryDsn: string, event: Event) => {
-  const components = dsnFromString(sentryDsn)
-  if (!components) {
-    throw new Error('invalid DSN provided')
-  }
+  const components = validateSentryDsn(sentryDsn)
 
   const envelope = createEventEnvelope(event)
   const endpoint = getEnvelopeEndpointWithUrlEncodedAuth(components)
@@ -38,4 +21,5 @@ export const validateSentryDsn = (sentryDsn: string) => {
   if (!components) {
     throw new Error('invalid DSN provided')
   }
+  return components
 }
