+++ packages/modules/sentry/tests/serialize-envelope.test.ts
@@ -1,0 +1,14 @@
+import { serializeEnvelope } from '../src/serialize-envelope'
+
+describe('serializeEnvelope', () => {
+  it('should concat buffers', () => {
+    expect(
+      serializeEnvelope([
+        // @ts-expect-error
+        new Uint8Array(),
+        // @ts-expect-error
+        [[{}, new Uint8Array()]],
+      ]),
+    ).toEqual(new Uint8Array([123, 125, 10, 123, 125, 10]))
+  })
+})
